+++ R/sf.R
@@ -137,12 +137,12 @@
 #'
 #' @examples
 #' sf_phrase("5 4 3 2 1", "1 3 3 3 1", "8*4 1", key = "b_")
-#' sf_phrase("654321 6s 12 1 21", "133211 355333 11 (13) (13)(13)", "4 4 8 8 4",
+#' sf_phrase("6 6 12 1 21", "133211 355333 11 (13) (13)(13)", "4 4 8 8 4",
 #'           key = "f")
-#' sfp("6s*2 1*4", "000232*2 2*4", "4 4 8*4", tuning = "dropD", key = "d")
+#' sfp("6*2 1*4", "000232*2 2*4", "4 4 8*4", tuning = "dropD", key = "d")
 #'
 #' # compare with single-argument input
-#' s <- "3s*5 53~*3 543*2 643"
+#' s <- "3*5 53~*3 543*2 643"
 #' f <- "987*2 775 553 335 77~*3 545 325 210"
 #' i <- "2*3 4. 16 4.*3 4*3"
 #' p1 <- sfp(s, f, i)
@@ -175,7 +175,7 @@ sf_phrase <- function(string, fret = NULL, info = NULL, key = "c",
       stop(paste("`info` must have the same number of timesteps as `string`",
                  "or a single value to repeat."), call. = FALSE)
   }
-  string <- .strsub(string)
+  string <- .sfp_infer_strings(fret, string)
   tuning <- .map_tuning(tuning)
   open_notes <- rev(strsplit(tuning, " ")[[1]])
   str_num <- rev(seq_along(open_notes))
@@ -190,7 +190,7 @@ sf_phrase <- function(string, fret = NULL, info = NULL, key = "c",
     if(x %in% rests | y %in% rests){
       if(x == y) return(x) else stop("Rest mismatch.", call. = FALSE)
     }
-    x <- as.integer(strsplit(x, "_")[[1]])
+    x <- as.integer(strsplit(x, "")[[1]])
     if(any(!x %in% str_num))
       stop("String number outside range inferred by tuning.", call. = FALSE)
     y <- trimws(gsub("(\\(\\d{1,2}\\))", " \\1 ", y))
@@ -213,7 +213,7 @@ sf_phrase <- function(string, fret = NULL, info = NULL, key = "c",
   })) %>%
     unlist() %>%
     paste(collapse = " ")
-  if(to_notes) return(notes)
+  if(to_notes) return(as_noteworthy(notes))
   phrase(notes, info, gsub("~", "", string), bar)
 }
 
@@ -232,6 +232,27 @@ sf_note <- function(...){
 #' @export
 #' @rdname sf_phrase
 sfn <- sf_note
+
+.sfp_infer_strings <- function(x, s){
+  size <- sapply(strsplit(.strsub(gsub("~", "", x)), "_"), length)
+  rests <- note_is_rest(x)
+  s_tie <- grepl("~", s)
+  if(any(s_tie)) s[s_tie] <- gsub("~", "", s[s_tie])
+  idx <- which(nchar(s) == 1 & size > 1)
+  if(length(idx)) s[idx] <- purrr::map2_chr(as.integer(s[idx]), size[idx], ~{
+    x <- paste(seq(.x, by = -1, length.out = .y), collapse = "")
+    if(grepl("[-0]", x)) stop("Invalid string number < 1.", call. = FALSE)
+    x
+  })
+  idx <- !rests & nchar(s) != 0
+  if(any(idx)){
+    if(any(nchar(s[idx]) != size[idx]))
+      stop("Number of strings and frets must match at each non-rest timestep.",
+           call. = FALSE)
+  }
+  if(any(s_tie)) s[s_tie] <- paste0(s[s_tie], "~")
+  s
+}
 
 .split_sfp_input <- function(x){
   x <- strsplit(x, " ")[[1]]
@@ -278,7 +299,6 @@ sfn <- sf_note
   muted <- .muted_index(fret)
   f <- function(string, fret, n, muted){
     if(n == 1 & fret %in% c("r", "s")) return(fret)
-    if(string == n & -1 %in% muted & n > 2) return(paste0(string, "s"))
     string <- seq(as.integer(string), by = -1, length.out = n)
     if(!(-1 %in% muted)) string <- string[-muted]
     paste0(string, collapse = "")